Wait for StatusNotifierWatcher's appearance
In some cases StatusNotifierWatcher may not be available when toggler
starts up. For example, when auto-starting on Sway. In such cases we
need to wait untill StatusNotifierWatcher service appears and only
register then. Also, the listener is not freed after successful
registration, so if the StatusNotifierWatcher changes for some reason -
toggler will re-register itself. Although I'm not sure if this can
really happen and if it's worth wasting DBus resources on.